New algorithm for managing the snapd state on classic

+++ tests/lib/state.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-SNAPD_STATE_PATH="$SPREAD_PATH/snapd-state"
-SNAPD_STATE_FILE="$SPREAD_PATH"/snapd-state.tar
+SNAPD_STATE_PATH="$GOPATH/snapd-state"
 
 # shellcheck source=tests/lib/dirs.sh
 . "$TESTSLIB/dirs.sh"
@@ -10,16 +9,22 @@ SNAPD_STATE_FILE="$SPREAD_PATH"/snapd-state.tar
 . "$TESTSLIB/boot.sh"
 
 delete_snapd_state() {
-    rm -f $SNAPD_STATE_FILE
     rm -rf $SNAPD_STATE_PATH
 }
 
 save_classic_state() {
-    escaped_snap_mount_dir=$1
-    snapd_env="/etc/environment /etc/systemd/system/snapd.service.d /etc/systemd/system/snapd.socket.d"
-    snap_confine_profiles="$(ls /etc/apparmor.d/snap.core.* || true)"
-    # shellcheck disable=SC2086
-    tar cf "$SNAPD_STATE_FILE" /var/lib/snapd "$SNAP_MOUNT_DIR" /etc/systemd/system/"$escaped_snap_mount_dir"-*core*.mount /etc/systemd/system/multi-user.target.wants/"$escaped_snap_mount_dir"-*core*.mount $snap_confine_profiles $snapd_env
+    local escaped_snap_mount_dir=$1
+
+    mkdir -p "$SNAPD_STATE_PATH" "$SNAPD_STATE_PATH"/snap-confine-profiles "$SNAPD_STATE_PATH"/system-units "$SNAPD_STATE_PATH"/multi-user-units
+
+    cp -rfp /var/lib/snapd "$SNAPD_STATE_PATH"/snapd-lib
+    cp -rf "$SNAP_MOUNT_DIR" "$SNAPD_STATE_PATH"/snap-mount-dir
+    cp -f /etc/systemd/system/"$escaped_snap_mount_dir"-*core*.mount "$SNAPD_STATE_PATH"/system-units
+    cp -f /etc/systemd/system/multi-user.target.wants/"$escaped_snap_mount_dir"-*core*.mount "$SNAPD_STATE_PATH"/multi-user-units
+    cp -f /etc/environment "$SNAPD_STATE_PATH"/environment
+    cp -rf /etc/systemd/system/snapd.service.d "$SNAPD_STATE_PATH"/snap.service.d
+    cp -rf /etc/systemd/system/snapd.socket.d "$SNAPD_STATE_PATH"/snap.socket.d
+    cp -rf /etc/apparmor.d/snap.core.* "$SNAPD_STATE_PATH"/snap-confine-profiles
 }
 
 restore_classic_state() {
@@ -27,18 +32,25 @@ restore_classic_state() {
     rm -rf /etc/systemd/system/snapd.service.d
     rm -rf /etc/systemd/system/snapd.socket.d
 
-    tar -C/ -xf "$SNAPD_STATE_FILE"
+    restore_snapd_lib
+    cp -rf "$SNAPD_STATE_PATH"/snap-mount-dir/* "$SNAP_MOUNT_DIR"
+    cp -f "$SNAPD_STATE_PATH"/system-units/* /etc/systemd/system
+    cp -f "$SNAPD_STATE_PATH"/multi-user-units/* /etc/systemd/system/multi-user.target.wants/
+    cp -f "$SNAPD_STATE_PATH"/environment /etc/environment
+    cp -rf "$SNAPD_STATE_PATH"/snap.service.d /etc/systemd/system/snapd.service.d
+    cp -rf "$SNAPD_STATE_PATH"/snap.socket.d /etc/systemd/system/snapd.socket.d
+    cp -rf "$SNAPD_STATE_PATH"/snap-confine-profiles/* /etc/apparmor.d
 }
 
 save_all_snap_state() {
     local boot_path="$(get_boot_path)"
 
-    mkdir -p "$SNAPD_STATE_PATH"
+    mkdir -p "$SNAPD_STATE_PATH" "$SNAPD_STATE_PATH"/system-units
 
     # Copy the state preserving the timestamps
-    cp -rfp /var/lib/snapd "$SNAPD_STATE_PATH/snapdlib"
-    cp -rf "$boot_path" "$SNAPD_STATE_PATH/boot"
-    cp -f /etc/systemd/system/snap-*core*.mount "$SNAPD_STATE_PATH"
+    cp -rfp /var/lib/snapd "$SNAPD_STATE_PATH"/snapd-lib
+    cp -rf "$boot_path" "$SNAPD_STATE_PATH"/boot
+    cp -f /etc/systemd/system/snap-*core*.mount "$SNAPD_STATE_PATH"/system-units
 }
 
 restore_all_snap_state() {
@@ -46,6 +58,12 @@ restore_all_snap_state() {
     # fully for tests that break it
     local boot_path="$(get_boot_path)"
 
+    restore_snapd_lib
+    cp -rf "$SNAPD_STATE_PATH"/boot/* "$boot_path"
+    cp -f "$SNAPD_STATE_PATH"/system-units/*  /etc/systemd/system
+}
+
+restore_snapd_lib() {
     mkdir -p /var/lib/snapd/snaps
     mkdir -p /var/lib/snapd/seed /var/lib/snapd/seed/snaps
 
@@ -55,42 +73,48 @@ restore_all_snap_state() {
     clean_seed_dir_state
 
     # Copy the whole state but the snaps and seed dirs
-    find "$SNAPD_STATE_PATH"/snapdlib/* -maxdepth 0 ! \( -name 'snaps' -o -name 'seed' \) -exec cp -rf {} /var/lib/snapd \;
+    find "$SNAPD_STATE_PATH"/snapd-lib/* -maxdepth 0 ! \( -name 'snaps' -o -name 'seed' \) -exec cp -rf {} /var/lib/snapd \;
     sync_snaps_dir_state
     sync_seed_dir_state
-
-    # Restore boot and mount points
-    cp -rf "$SNAPD_STATE_PATH"/boot/* "$boot_path"
-    cp -f "$SNAPD_STATE_PATH"/snap-*core*.mount  /etc/systemd/system
 }
 
 clean_snaps_dir_state() {
-    find /var/lib/snapd/snaps/* -maxdepth 0 ! -name '*.snap' -exec rm -rf {} +
+    if [ ! -z "$(ls -A /var/lib/snapd/snaps)" ]; then
+        find /var/lib/snapd/snaps/* -maxdepth 0 ! -name '*.snap' -exec rm -rf {} +
+    fi
 }
 
 clean_seed_dir_state() {
     find /var/lib/snapd/seed/* -maxdepth 0 ! -name 'snaps' -exec rm -rf {} +
-    find /var/lib/snapd/seed/snaps/* -maxdepth 0 ! -name '*.snap' -exec rm -rf {} +
+    if [ ! -z "$(ls -A /var/lib/snapd/seed/snaps)" ]; then
+        find /var/lib/snapd/seed/snaps/* -maxdepth 0 ! -name '*.snap' -exec rm -rf {} +
+    fi
 }
 
 sync_snaps_dir_state() {
-    find "$SNAPD_STATE_PATH"/snapdlib/snaps/* -maxdepth 0 ! -name '*.snap' -exec cp -rf {} /var/lib/snapd/snaps \;
-    sync_snaps "$SNAPD_STATE_PATH"/snapdlib/snaps /var/lib/snapd/snaps
+    if [ ! -z "$(ls -A "$SNAPD_STATE_PATH"/snapd-lib/snaps)" ]; then
+        find "$SNAPD_STATE_PATH"/snapd-lib/snaps/* -maxdepth 0 ! -name '*.snap' -exec cp -rf {} /var/lib/snapd/snaps \;
+    fi
+    sync_snaps "$SNAPD_STATE_PATH"/snapd-lib/snaps /var/lib/snapd/snaps
 }
 
 sync_seed_dir_state() {
-    find "$SNAPD_STATE_PATH"/snapdlib/seed/* -maxdepth 0 ! -name 'snaps' -exec cp -rf {} /var/lib/snapd/seed \;
-    find "$SNAPD_STATE_PATH"/snapdlib/seed/snaps/* -maxdepth 0 ! -name '*.snap' -exec cp -rf {} /var/lib/snapd/seed/snaps \;
-    sync_snaps "$SNAPD_STATE_PATH"/snapdlib/seed/snaps /var/lib/snapd/seed/snaps
+    if [ ! -z "$(ls -A "$SNAPD_STATE_PATH"/snapd-lib/seed)" ]; then
+        find "$SNAPD_STATE_PATH"/snapd-lib/seed/* -maxdepth 0 ! -name 'snaps' -exec cp -rf {} /var/lib/snapd/seed \;
+    fi
+    if [ ! -z "$(ls -A "$SNAPD_STATE_PATH"/snapd-lib/seed/snaps)" ]; then
+        find "$SNAPD_STATE_PATH"/snapd-lib/seed/snaps/* -maxdepth 0 ! -name '*.snap' -exec cp -rf {} /var/lib/snapd/seed/snaps \;
+    fi
+    sync_snaps "$SNAPD_STATE_PATH"/snapd-lib/seed/snaps /var/lib/snapd/seed/snaps
 }
 
 sync_snaps() {
-    SOURCE=$1
-    TARGET=$2
+    local SOURCE=$1
+    local TARGET=$2
 
     if ! [ -d "$SOURCE" ]; then
-        echo "Source directory does not exist $SOURCE"
-        exit 1
+        rm -rf "$TARGET"
+        return
     elif ! [ -d "$TARGET" ]; then
         mkdir -p "$TARGET"
     fi
